requirements.txt
fastapi[standard]>=0.116.1
numpy>=2.3.2
openmeteo-requests>=1.6.0
pandas>=2.3.1
psycopg2>=2.9.10
requests-cache>=1.2.1
retry-requests>=2.0.0
sqlalchemy>=2.0.42